Added rest repo for all components

+++ src/main/java/com/computerservice/repository/role/RoleEntityRepository.java
@@ -3,6 +3,8 @@ package com.computerservice.repository.role;
 import com.computerservice.entity.role.RoleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RoleEntityRepository extends JpaRepository<RoleEntity, Integer> {
+import java.math.BigInteger;
+
+public interface RoleEntityRepository extends JpaRepository<RoleEntity, BigInteger> {
     RoleEntity findByName(String name);
 }
